Refactored src to commons/ and jophiel/
