+++ config/webpack.config.js
@@ -24,7 +24,7 @@ const webpackConfig = {
 // ------------------------------------
 // Entry Points
 // ------------------------------------
-const APP_ENTRY = project.paths.client('main.js')
+const APP_ENTRY = project.paths.client('jophiel', 'main.js')
 
 webpackConfig.entry = {
   app : __DEV__
