+++ src/commons/components/Header/Header.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import './Header.scss'
-import Logo from '../../assets/images/logo.png'
+import Logo from 'assets/images/logo.png'
 
 export const Header = () => (
   <header>
