+++ src/jophiel/components/UserProfileWidget/avatar-default.png
