+++ src/jophiel/containers/UserProfileWidgetContainer.js
