+++ src/jophiel/layouts/CoreLayout/CoreLayout.js
@@ -1,12 +1,12 @@
 import React from 'react'
-import Header from '../../components/Header'
-import Footer from '../../components/Footer'
-import Breadcrumbs from '../../components/Breadcrumbs'
+import Header from 'commons/components/Header'
+import Footer from 'commons/components/Footer'
+import Breadcrumbs from 'commons/components/Breadcrumbs'
 import TwoColumnLayout from '../TwoColumnLayout'
-import UserProfileWidgetContainer from '../../containers/UserProfileWidgetContainer'
+import UserProfileWidgetContainer from 'jophiel/containers/UserProfileWidgetContainer'
 import ContentLayout from '../ContentLayout'
 import './CoreLayout.scss'
-import '../../styles/core.scss'
+import 'styles/core.scss'
 
 export class CoreLayout extends React.Component {
   render () {
