+++ src/jophiel/layouts/CoreLayout/CoreLayout.scss
