+++ src/jophiel/layouts/TwoColumnLayout/index.js
